feat(bundles): lazy-load heavy chart components with next/dynamic
Converts EarningsLineChart, OrdersBarChart, UsersGrowthChart, and CategoryPieChart
to use next/dynamic with ssr: false. Charts are deferred until client-side render,
reducing initial bundle size for dashboard pages.

Pages optimized:
- Dashboard overview page
- Earnings page
- Admin analytics page

Shows skeleton placeholders while charts load.

Resolves #471
